modelの切り出し
