+++ lib/api/auth_api.dart
@@ -1,10 +1,7 @@
 import 'dart:convert';
 
-import 'package:flutter/material.dart';
-import 'package:json_annotation/json_annotation.dart';
 import 'package:mobile/api/api_client.dart';
-
-part 'auth_api.g.dart';
+import 'package:mobile/model/auth.dart';
 
 abstract class AuthApi {
   Future<String> signUp(SignUpRequestBody body);
@@ -34,58 +31,7 @@ class DefaultAuthApi extends AuthApi {
       '/users/sign-up',
       body: json.encode(body),
     );
-    return AuthResponse.fromJson(
-            jsonDecode(response.body) as Map<String, dynamic>)
+    return Auth.fromJson(jsonDecode(response.body) as Map<String, dynamic>)
         .token;
   }
-}
-
-@JsonSerializable(includeIfNull: false)
-class SignUpRequestBody {
-  SignUpRequestBody({
-    @required this.name,
-    @required this.email,
-    @required this.password,
-  });
-
-  String name;
-  String email;
-  String password;
-
-  factory SignUpRequestBody.fromJson(Map<String, dynamic> json) =>
-      _$SignUpRequestBodyFromJson(json);
-
-  Map<String, dynamic> toJson() => _$SignUpRequestBodyToJson(this);
-}
-
-@JsonSerializable(includeIfNull: false)
-class SignInRequestBody {
-  SignInRequestBody({
-    @required this.email,
-    @required this.password,
-  });
-
-  String email;
-  String password;
-
-  factory SignInRequestBody.fromJson(Map<String, dynamic> json) =>
-      _$SignInRequestBodyFromJson(json);
-
-  Map<String, dynamic> toJson() => _$SignInRequestBodyToJson(this);
-}
-
-@JsonSerializable(includeIfNull: false)
-class AuthResponse {
-  AuthResponse({
-    @required this.token,
-    @required this.userId,
-  });
-
-  String token;
-  String userId;
-
-  factory AuthResponse.fromJson(Map<String, dynamic> json) =>
-      _$AuthResponseFromJson(json);
-
-  Map<String, dynamic> toJson() => _$AuthResponseToJson(this);
 }
